+++ src/runtime/Classes/Utilities.m
@@ -72,7 +72,7 @@ uint32_t convertInt32ToUInt32(int32_t v) {
 }
 
 
-int64_t convertUInt32ToInt32(uint32_t v) {
+int32_t convertUInt32ToInt32(uint32_t v) {
   union { int32_t i; uint32_t u; } u;
   u.u = v;
   return u.i;
